requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.95
botocore==1.34.95
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
cleo==2.1.0
crashtest==0.4.1
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-storages==1.14.2
django-taggit==5.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
drf-yasg==1.21.7
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.4
frozenlist==1.4.1
iamport-rest-client==0.9.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jmespath==1.0.1
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
oauthlib==3.2.2
packaging==24.0
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
requests==2.26.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.4
trove-classifiers==2024.4.10
twilio==9.0.5
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.25.3
xattr==1.1.0
yarl==1.9.4
zipp==3.18.1
